Updated io wrapping to be buildable, no functionality
