+++ Makefile
@@ -1,6 +1,6 @@
 CC = riscv64-unknown-linux-gnu-gcc
 CFLAGS = -Wall -Werror -fPIC
-SRCS = page_fault.c interrupt.c printf.c syscall.c string.c linux_wrap.c
+SRCS = page_fault.c interrupt.c printf.c syscall.c string.c linux_wrap.c io_wrap.c
 ASM_SRCS = entry.S
 RUNTIME = eyrie-rt
 LINK = riscv64-unknown-linux-gnu-ld
